LayoutTests/imported/w3c/web-platform-tests/navigation-timing/navigation-type-post-backforward-expected.txt



FAIL Navigation type after posting and navigating away and back should be back_forward. assert_equals: expected "back_forward" but got "navigate"
